use hex notation
Another trivial editorial suggestion. On initial read I paused for a second to consider why the code point was listed as 21 above yet 15 here. Obviously it's in hex; adding the "0x" when writing inline makes it explicit. I doubt any implementer would make a mistake here, but it makes it more clear nonetheless.